+++ CHANGELOG.md
@@ -6,6 +6,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- `Removable` controller from [@CiTroNaK](https://github.com/CiTroNaK)
+
 ## [1.0.0] - 2020-10-15
 
 ### Added